// src/types/contactForm.ts
export type ContactForm = {
    title: string;
    description: string;
    name: string;
    email: string;
    object: string;
    message: string;
    errors_name: string;
    errors_email: string;
    errors_object: string;
    errors_message: string;
    global_error: string;
    submit: string;
    success_message: string;
  };